Remove graphql_types crate
