--- backend/graphql_types/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "graphql_types"
-version = "0.1.0"
-edition = "2021"
-
-
-[dependencies]
-async-graphql = { workspace = true }
-thiserror = { workspace = true }
-uuid = { workspace = true }
