--- backend/graphql_types/README.md
@@ -1,3 +1,0 @@
-# GraphQL Types
-
-A collection of GraphQL types which are common across services.
