+++ backend/pin_packing/Cargo.toml
@@ -11,7 +11,6 @@ clap = { workspace = true }
 chrono = { workspace = true }
 dotenvy = { workspace = true }
 graphql_endpoints = { path = "../graphql_endpoints" }
-graphql_types = { path = "../graphql_types" }
 opa_client = { path = "../opa_client", features = ["graphql"] }
 sea-orm = { workspace = true, features = ["sqlx-postgres"] }
 sea-orm-migration = { workspace = true }
