+++ backend/pin_packing/src/resolvers/mod.rs
@@ -5,3 +5,13 @@ pub mod pin_library;
 pub mod pin_mount;
 pub mod puck_library;
 pub mod puck_mount;
+
+use async_graphql::{InputObject, SimpleObject};
+use uuid::Uuid;
+
+#[derive(Debug, Clone, SimpleObject, InputObject)]
+#[graphql(input_name = "WellInput")]
+pub struct Well {
+    pub plate: Uuid,
+    pub well: i16,
+}
